% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods.R
\name{get_changelogs}
\alias{get_changelogs}
\title{Get changelogs}
\usage{
get_changelogs(game_domain_name, mod_id)
}
\arguments{
\item{game_domain_name}{A character. Game domain to get the recent mod updates for.}

\item{mod_id}{A numeric. ID of the mod to get changelogs for.}
}
\value{

}
\description{
Get changelogs
}
\examples{
\dontrun{
library(nexusmodsr)

nexus_auth()

changelogs <- get_changelogs("Stardew Valley", 2400)
}
}
\references{
https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/Mods/get_v1_games_game_domain_mods_mod_id_changelogs.json
}
\seealso{
Other Mods: 
\code{\link{get_latest_added_mods}()},
\code{\link{get_latest_updated_mods}()},
\code{\link{get_mod}()},
\code{\link{get_search_md5_hash}()},
\code{\link{get_trending_mods}()},
\code{\link{mods_updated}()},
\code{\link{post_endorse_mod}()}
}
\concept{Mods}
