% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods.R
\name{get_mod}
\alias{get_mod}
\title{Get mod}
\usage{
get_mod(game_domain_name, mod_id)
}
\arguments{
\item{game_domain_name}{A character. Game domain to get the recent mod updates for.}

\item{mod_id}{A numeric. ID of the mod to get changelogs for.}
}
\value{

}
\description{
Retrieve specified mod, from a specified game. Cached for 5 minutes.
}
\examples{
\dontrun{
library(nexusmodsr)

nexus_auth()

mod <- get_mod("Stardew Valley",  2400)
}
}
\references{
https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/Mods/get_v1_games_game_domain_name_mods_id.json
}
\seealso{
Other Mods: 
\code{\link{get_changelogs}()},
\code{\link{get_latest_added_mods}()},
\code{\link{get_latest_updated_mods}()},
\code{\link{get_search_md5_hash}()},
\code{\link{get_trending_mods}()},
\code{\link{mods_updated}()},
\code{\link{post_endorse_mod}()}
}
\concept{Mods}
