% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods.R
\name{get_search_md5_hash}
\alias{get_search_md5_hash}
\title{Search MD5 Hash}
\usage{
get_search_md5_hash(game_domain_name, md5_hash)
}
\arguments{
\item{game_domain_name}{A character. Game domain to get the recent mod updates for.}

\item{md5_hash}{A character. I'll be honest, I don't know what this means.}
}
\value{

}
\description{
Looks up a file MD5 file hash
}
\references{
https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/Mods/get_v1_games_game_domain_name_mods_md5_search_md5_hash.json
}
\seealso{
Other Mods: 
\code{\link{get_changelogs}()},
\code{\link{get_latest_added_mods}()},
\code{\link{get_latest_updated_mods}()},
\code{\link{get_mod}()},
\code{\link{get_trending_mods}()},
\code{\link{mods_updated}()},
\code{\link{post_endorse_mod}()}
}
\concept{Mods}
