% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods.R
\name{get_trending_mods}
\alias{get_trending_mods}
\title{Get trendings mods}
\usage{
get_trending_mods(game_domain_name)
}
\arguments{
\item{game_domain_name}{A character. Game domain to get the recent mod updates for.}
}
\value{

}
\description{
Retrieve 10 trending mods for a specified game
}
\examples{
\dontrun{
library(nexusmodsr)

nexus_auth()

trending_mods <- get_trending_mods("Stardew Valley")
}
}
\references{
https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/Mods/get_v1_games_game_domain_mods_trending.json
}
\seealso{
Other Mods: 
\code{\link{get_changelogs}()},
\code{\link{get_latest_added_mods}()},
\code{\link{get_latest_updated_mods}()},
\code{\link{get_mod}()},
\code{\link{get_search_md5_hash}()},
\code{\link{mods_updated}()},
\code{\link{post_endorse_mod}()}
}
\concept{Mods}
