% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods.R
\name{mods_updated}
\alias{mods_updated}
\title{Get recently updated mods}
\usage{
mods_updated(game_domain_name, period)
}
\arguments{
\item{game_domain_name}{A character. Game domain to get the recent mod updates for.}

\item{period}{A character. Requests the wanted period. Format is "1d", "1w", and "1m" for day, week and month.}
}
\value{
A tibble with \code{mod_id}, \code{lastest_file_update}, and \code{lastest_mod_activity}
}
\description{
Get recently updated mods
}
\examples{
\dontrun{
library(nexusmodsr)

nexus_auth()

recent_updates <- mods_updated("Stardew Valley", "1w")
}
}
\references{
https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/Mods/get_v1_games_game_domain_mods_updates.json
}
\seealso{
Other Mods: 
\code{\link{get_changelogs}()},
\code{\link{get_latest_added_mods}()},
\code{\link{get_latest_updated_mods}()},
\code{\link{get_mod}()},
\code{\link{get_search_md5_hash}()},
\code{\link{get_trending_mods}()},
\code{\link{post_endorse_mod}()}
}
\concept{Mods}
