% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexus_auth.R
\name{nexus_auth}
\alias{nexus_auth}
\title{Add API Key to all requests}
\usage{
nexus_auth(nexus_mods_key = Sys.getenv("NEXUSMODS_KEY"))
}
\arguments{
\item{nexus_mods_key}{A character. Nexus Mods Personal API Key.}
}
\value{
Invisible authorization data.
}
\description{
Add API Key to all requests
}
\examples{
\dontrun{
library(nexusmodsr)

nexus_mods_key()
}
}
