% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexusmodsr-package.R
\docType{package}
\name{nexusmodsr-package}
\alias{nexusmodsr}
\alias{nexusmodsr-package}
\title{nexusmodsr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
