% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods.R
\name{post_abstain_mod}
\alias{post_abstain_mod}
\title{Abstain a mod}
\usage{
post_abstain_mod(game_domain_name, mod_id, version)
}
\arguments{
\item{game_domain_name}{A character. Game domain to get the recent mod updates for.}

\item{mod_id}{A numeric. ID of the mod to get changelogs for.}

\item{version}{A character. Mod version.}
}
\value{

}
\description{
Abstain from endorsing a mod
}
\examples{
\dontrun{
library(nexusmodsr)

nexus_auth()

abstain <- post_abstain_mod("Stardew Valley",  2400, "3.11.0")
}
}
\references{
https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/Mods/post_v1_games_game_domain_name_mods_id_abstain.json
}
\concept{mod}
