% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods.R
\name{post_endorse_mod}
\alias{post_endorse_mod}
\title{Endorse a mod}
\usage{
post_endorse_mod(game_domain_name, mod_id, version)
}
\arguments{
\item{game_domain_name}{A character. Game domain to get the recent mod updates for.}

\item{mod_id}{A numeric. ID of the mod to get changelogs for.}

\item{version}{A character. Mod version.}
}
\value{

}
\description{
Endorse a mod
}
\examples{
\dontrun{
library(nexusmodsr)

nexus_auth()

endorse <- post_endorse_mod("Stardew Valley",  2400, "3.11.0")
}
}
\references{
https://app.swaggerhub.com/apis-docs/NexusMods/nexus-mods_public_api_params_in_form_data/1.0#/Mods/post_v1_games_game_domain_name_mods_id_endorse.json
}
\seealso{
Other Mods: 
\code{\link{get_changelogs}()},
\code{\link{get_latest_added_mods}()},
\code{\link{get_latest_updated_mods}()},
\code{\link{get_mod}()},
\code{\link{get_search_md5_hash}()},
\code{\link{get_trending_mods}()},
\code{\link{mods_updated}()}
}
\concept{Mods}
